% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_functions.R
\name{sample_highed}
\alias{sample_highed}
\alias{p_eddem}
\alias{sample_eddem}
\title{High News Interest + Democrat Oversample}
\source{
Inspired by Lauren Kennedy's code in \href{https://mc-stan.org/rstanarm/articles/mrp.html}{rstanarm}
}
\usage{
sample_highed(data, n)

p_eddem(data)

sample_eddem(data, n)
}
\arguments{
\item{data}{A tibble with the columns race, educ, newsint, as coded by the CCES}

\item{n}{Sample size to sample}
}
\value{
A vector of propensity scores ranging from 0 to 1
}
\description{
Propensity Score that oversamples high education, high-news
interest, and Democratic people
}
\details{
We model the true propensity score as being a separable function of
race, education, news interest, and party.
}
\examples{
p_eddem(sample_n(pop_cces, 100))

}
