% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svysim.R
\name{report_pid_dem}
\alias{report_pid_dem}
\title{Get a sample estimate}
\usage{
report_pid_dem(tbl)
}
\arguments{
\item{tbl}{A dataframe with the variable pid3_leaner}
}
\value{
A scalar, in this case the proportion of Democrats (including leaners)
}
\description{
Get a sample estimate
}
