% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_functions.R
\name{samp_srs}
\alias{samp_srs}
\title{Simple Random Sample}
\usage{
samp_srs(data, n)
}
\arguments{
\item{data}{A tibble with the columns race, educ, newsint, as coded by the CCES}

\item{n}{Sample size to sample}
}
\description{
Simple Random Sample
}
