% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svysim.R
\name{samp_with}
\alias{samp_with}
\title{Sample with custom function}
\usage{
samp_with(population, sampling_f, n)
}
\arguments{
\item{population}{An object defined by \link{DeclareDesign::declare_design}. See
help page here.}

\item{sampling_f}{A sampling function that can be passed as a handler to
\code{declare_sampling}.}

\item{n}{the sample size to draw}
}
\description{
Sample with custom function
}
\examples{

# Population Value
popn <- declare_population(pop_cces)
stat_dem(pop_cces)

# SRS
samp0  <- samp_with(popn, sample_srs, n = 1000))
stat_dem(samp0)

# Oversample Higher-ed
samp1  <- samp_with(popn, sample_highed, n = 1000)
stat_dem(samp1)

}
