% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_functions.R
\name{sample_pscore}
\alias{sample_pscore}
\title{Sample with custom propensity score variable}
\usage{
sample_pscore(data, varname, n)
}
\arguments{
\item{varname}{column name which contains the propensity score, unquoted.}

\item{n}{sample size}
}
\description{
Sample with custom propensity score variable
}
