% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_statistic.R
\name{stat_dem}
\alias{stat_dem}
\alias{stat_Z}
\alias{stat_Y}
\title{Get a sample estimate}
\usage{
stat_dem(tbl)

stat_Z(tbl)

stat_Y(tbl)
}
\arguments{
\item{tbl}{A dataframe with the variable pid3_leaner}
}
\value{
A scalar, in this case the proportion of Democrats (including leaners)
}
\description{
\code{stat_dem} computes the mean Democrats, \code{stat_Z} computes the sample proportion
of the simulated oucome \code{Z}, \code{stat_Y} the mean of the simulated continuous variable
\code{Y}.
}
