% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc_pop-cces.R
\docType{data}
\name{urban_CDs}
\alias{urban_CDs}
\title{List of Urban Congressional Districts}
\format{
An object of class \code{character} of length 34.
}
\usage{
urban_CDs
}
\description{
Districts that are classified as "pure urban" by CityLab's 2018 classification
}
\keyword{datasets}
