% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svysim.R
\name{p_highed}
\alias{p_highed}
\alias{sample_highed}
\title{Propensity Score Simulation}
\usage{
p_highed(data)

sample_highed(data, n)
}
\arguments{
\item{data}{A tibble with the columns race, educ, newsint, as coded by the CCES}

\item{n}{Sample size to sample}
}
\value{
A vector of propensity scores ranging from 0 to 1
}
\description{
Propensity Score that oversamples high education and high-news interest people
}
