% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_functions.R
\name{p_highed}
\alias{p_highed}
\alias{samp_highed}
\title{High News Interest Oversample}
\source{
Inspired by Lauren Kennedy's code in \href{https://mc-stan.org/rstanarm/articles/mrp.html}{rstanarm}
}
\usage{
p_highed(data)

samp_highed(data, n)
}
\arguments{
\item{data}{A tibble with the columns race, educ, newsint, as coded by the CCES}

\item{n}{Sample size to sample}
}
\value{
A vector of propensity scores ranging from 0 to 1
}
\description{
Propensity Score that oversamples high education and high-news interest people
}
\details{
We model the true propensity score as being a separable function of
race, education, and news interest. There is no explicit correlation between
selection and the outcome (say, partisanship), so Meng's rho would be non-zero
only so far as the covariates correlates with race, education, and news interest.
}
\examples{
p_highed(samp_n(pop_cces, 100))

}
